Changed access to database to JDBC from JPA by the use of Datasource

+++ src/main/java/com/project/stone/user/entity/User.java
@@ -19,7 +19,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Integer id;
 
     @Column(name = "username", nullable = false, length = 50)
     private String username;
@@ -31,7 +31,7 @@ public class User {
     private String password;
 
     @Column(name = "created_at")
-    private Integer createdAt;
+    private Long createdAt;
